Refactor Consulardoc to Demand: Remove Consulardoc model, migration, and views; create Demand model, migration, and corresponding views. Update routes and controllers to handle demands instead of consulardocs. Adjust data types and relationships in migrations and models. Modify frontend forms and tables for demands management.

+++ database/migrations/2026_03_15_031040_create_documents_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('icone', 50);
             $table->string('libelle', 50);
             $table->decimal('amount', 10, 0);
-            $table->integer('day');
+            $table->tinyInteger('day');
             $table->text('description')->nullable();
             $table->tinyInteger('position')->default('0');
             $table->tinyInteger('status')->default('1');
